Convert to use mixins + basic layout
