+++ dis_snek/event_processors/guild_events.py
@@ -1,0 +1,39 @@
+import logging
+
+from dis_snek.const import logger_name
+from dis_snek.event_processors._template import EventMixinTemplate
+from dis_snek.models import listen, events
+from dis_snek.models.events import RawGatewayEvent
+
+log = logging.getLogger(logger_name)
+
+
+class GuildEvents(EventMixinTemplate):
+    @listen()
+    async def _on_raw_guild_create(self, event: RawGatewayEvent) -> None:
+        """
+        Automatically cache a guild upon GUILD_CREATE event from gateway.
+
+        Args:
+            event: raw guild create event
+        """
+        guild = self.cache.place_guild_data(event.data)
+        self._guild_event.set()
+
+        self.dispatch(events.GuildCreate(guild))
+
+    @listen()
+    async def _on_raw_guild_update(self, event: RawGatewayEvent) -> None:
+        ...
+
+    @listen()
+    async def _on_raw_guild_delete(self, event: RawGatewayEvent) -> None:
+        ...
+
+    @listen()
+    async def _on_raw_guild_ban_add(self, event: RawGatewayEvent) -> None:
+        ...
+
+    @listen()
+    async def _on_raw_guild_ban_remove(self, event: RawGatewayEvent) -> None:
+        ...
