+++ dis_snek/event_processors/thread_events.py
@@ -1,0 +1,22 @@
+import logging
+
+from dis_snek.const import logger_name
+from dis_snek.event_processors._template import EventMixinTemplate
+from dis_snek.models import listen, events
+from dis_snek.models.events import RawGatewayEvent
+
+log = logging.getLogger(logger_name)
+
+
+class ThreadEvents(EventMixinTemplate):
+    @listen()
+    async def _on_raw_thread_create(self, event: RawGatewayEvent) -> None:
+        ...
+
+    @listen()
+    async def _on_raw_thread_update(self, event: RawGatewayEvent) -> None:
+        ...
+
+    @listen()
+    async def _on_raw_thread_delete(self, event: RawGatewayEvent) -> None:
+        ...
